Fix issue preventing avr-size from running on Linux
There is a bug in the patch that adds the `--format=avr` flag in the
avr-size command that leads to the command failing. This PR switches to using the
short flag (`-A`) for the avr output format as a work around.

The bug already has an open bug report but change made here will still work after
the fix has been implemented so there won't be a need to change going forward.