Add ability to detect changes to program settings
The `00-FlashUnix.sh` script will now generate an md5sum of the programs
.c file after a build is successful.

When both the md5sum file and the hex file exist for a program, the
script will check the current md5sum of the programs .c file and compare
against the checksum stored in the .md5sum file. If they do not match,
the script will give you the option to rebuild your hex file.