Add LED Support for Pro Micro
Added LED support for the Pro Micro board

Simplified the hardware detection logic by moving away from custom
defined `BOARD_TYPE`.

The LUFA library provides the `BOARD` macro already which allows for compiling based on a specific board. This facilitates having LED support on boards that use the same MCU but with different LED pins.

With this change, all of the various boards LED logic can be defined
under the same `setup_leds` and `onboard_leds` functions rather than
having to define them in multiple places (ie. the UNO logic was able to
be moved to the same function declarations as the rest and no longer
requires it's own definitions)